packages/natsy/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = ["pytest", "pytest-asyncio"]

[project]
name = "natsy"
description = "Simple helper functions for utilizing nats-py"
version = "0.0.1"
dependencies = ["nats-py", "loguru", "fastavro"]

[tool.pytest.ini_options]
pythonpath = ["src"]
